refactor: domain 하위로 폴더구조 변경
