+++ src/main/java/com/baedalping/delivery/domain/product/repository/ProductRepository.java
@@ -1,0 +1,9 @@
+package com.baedalping.delivery.domain.product.repository;
+
+import com.baedalping.delivery.domain.product.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface ProductRepository extends JpaRepository<Product, UUID> {
+}
