+++ src/main/java/com/baedalping/delivery/domain/storeCategory/dto/StoreCategoryCreateResponseDto.java
@@ -1,0 +1,18 @@
+package com.baedalping.delivery.domain.storeCategory.dto;
+
+import com.baedalping.delivery.domain.storeCategory.entity.StoreCategory;
+import java.util.UUID;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class StoreCategoryCreateResponseDto {
+  private UUID storeCategoryId;
+  private String storeCategoryName;
+
+  public StoreCategoryCreateResponseDto(StoreCategory storeCategory){
+    this.storeCategoryId = storeCategory.getStoreCategoryId();
+    this.storeCategoryName = storeCategory.getStoreCategoryName();
+  }
+}
