+++ src/main/java/com/baedalping/delivery/domain/storeCategory/repository/StoreCategoryRepository.java
@@ -1,0 +1,10 @@
+package com.baedalping.delivery.domain.storeCategory.repository;
+
+import com.baedalping.delivery.domain.storeCategory.entity.StoreCategory;
+import java.util.Optional;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StoreCategoryRepository extends JpaRepository<StoreCategory, UUID> {
+  Optional<StoreCategory> findByStoreCategoryName(String storeCategoryName);
+}
