--- src/main/java/com/baedalping/delivery/product/ProductRepository.java
@@ -1,8 +1,0 @@
-package com.baedalping.delivery.product;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface ProductRepository extends JpaRepository<Product, UUID> {
-}
