--- src/main/java/com/baedalping/delivery/productCategory/ProductCategoryRepository.java
@@ -1,10 +1,0 @@
-package com.baedalping.delivery.productCategory;
-
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface ProductCategoryRepository extends JpaRepository<ProductCategory, UUID> {
-  Optional<ProductCategory> findByProductCategoryName(String producCategoryName);
-}
