--- src/main/java/com/baedalping/delivery/store/StoreRepository.java
@@ -1,8 +1,0 @@
-package com.baedalping.delivery.store;
-
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StoreRepository extends JpaRepository<Store, UUID> {
-
-}
