--- src/main/java/com/baedalping/delivery/storeCategory/StoreCategoryCreateRequestDto.java
@@ -1,15 +1,0 @@
-package com.baedalping.delivery.storeCategory;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class StoreCategoryCreateRequestDto {
-  @NotBlank(message = "상품 분류명을 입력해주세요.")
-  private String storeCategoryName;
-}
