--- src/main/java/com/baedalping/delivery/storeCategory/StoreCategoryRepository.java
@@ -1,9 +1,0 @@
-package com.baedalping.delivery.storeCategory;
-
-import java.util.Optional;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StoreCategoryRepository extends JpaRepository<StoreCategory, UUID> {
-  Optional<StoreCategory> findByStoreCategoryName(String storeCategoryName);
-}
